@@ -1,17 +1,17 @@
 {
   "name": "wp-react-app",
   "private": true,
-  "version": "1.0.3",
+  "version": "1.1.0",
   "homepage": "http://localhost/wp-react-app/build",
   "dependencies": {
-    "@wordpress/block-library": "^6.0.2",
+    "@wordpress/block-library": "^7.5.0",
     "bootstrap": "^5.1.3",
-    "node-sass": "^6.0.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.1",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^4.0.3",
+    "sass": "^1.51.0",
+    "react": "^18.1.0",
+    "react-bootstrap": "^2.4.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "wpapi": "^1.2.2"
   },
   "scripts": {
