@@ -1,7 +1,6 @@
 import React from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
 import { Navbar, Nav, NavItem, NavLink, Dropdown } from 'react-bootstrap';
-import { Link } from 'react-router-dom';
 
 import '@wordpress/block-library/build-style/style.css';
 import 'bootstrap/scss/bootstrap.scss';
@@ -10,47 +9,12 @@ import logo from './logo.svg';
 
 import data from './_setup.json';
 
-import PageView from './pageView';
-import PostList from './postList';
-import PostView from './postView';
-
-import Search from './autocomplete';
-
-
-/**
- * Create all required page templates
- * 
- * @param {*} p 
- */
-
-const template = {};
-
-// Home
-template['home'] = () => (
-  <div>
-    <PageView page={ data.url.slug_home } />
-  </div>
-);
-
-// Post List
-template['posts'] = () => (
-  <div>
-    <div className="page posts">
-      <h1>Posts</h1>
-
-      <PostList />
-    </div>
-  </div>
-);
-
-// 404
-template['404'] = ({ location }) => (
-  <div>
-    <div className="page">
-      <h3>404. Not Found. <code>{ location.pathname }</code></h3>
-    </div>
-  </div>
-);
+import PageHome from './components/pageHome';
+import PageNotFound from './components/pageNotFound';
+import PageView from './components/pageView';
+import PostList from './components/postList';
+import PostView from './components/postView';
+import Search from './components/autocomplete';
 
 
 /**
@@ -106,14 +70,13 @@ const App = () => (
         </Navbar.Collapse>
       </Navbar>
 
-      <Switch>
-        <Route exact path={ '/' } component={ template[ 'home' ] } />
-        <Route exact path={ '/signin/' } component={ template[ 'signin' ] } />
-        <Route exact path={ '/' + data.url.slug_posts + '/:post' } component={ PostView } />
-        <Route exact path={ '/' + data.url.slug_posts + '/' } component={ PostList } />
-        <Route exact path={ '/:page' } component={ PageView } />
-        <Route component={ template[ '404' ] } />
-      </Switch>
+      <Routes>
+        <Route index element={ <PageHome /> } />
+        <Route path={ '/' + data.url.slug_posts + '/' } element={ <PostList /> } />
+        <Route path={ '/' + data.url.slug_posts + '/:post' } element={ <PostView /> } />
+        <Route path={ '/:page' } element={ <PageView /> } />
+        <Route path="*" element={ <PageNotFound /> } />
+      </Routes>
     </div>
   </Router>
 );
